Rename service/ modules to human-readable names, document service/ in CLAUDE.md
Executes step 0 of docs/design-plans/2026-08-24-csv-deliverable-and-component-names.md:
one noun per component, run-* prefix for anything executable.

  selection.js        -> article-picker.js
  pipeline.js          -> claim-extractor.js
  verify.js            -> verifier.js
  assemble.js          -> finding-builder.js
  findings.js          -> findings-store.js
  replay.js            -> run-replay.js
  extract-articles.js  -> run-extract.js
  select-articles.js   -> run-pick.js

Test files renamed to match. All internal imports, HELP_TEXT usage strings,
and header-comment cross-references updated; 694/694 tests pass.

Adds a "Batch Pipeline (service/)" section to CLAUDE.md, previously the only
undocumented top-level directory, with the component table and what's not
yet built (no runner joins all six stages). The two prior design docs that
predate the rename get a pointer note rather than a rewrite, per
docs/design-plans/README.md's rule against silently rewriting a design to
match the code.
