+++ CLAUDE.md
@@ -39,6 +39,7 @@ benchmark/
   analysis_v1.json               # Frozen v1 analysis snapshot
   analysis_v3.json               # Frozen v3 analysis snapshot
 Benchmarking_data_Citations.csv  # Source ground truth data (Dataset version + WMF override columns)
+service/                         # Toolforge batch pipeline: select -> extract -> fetch -> verify -> store (see Batch Pipeline below)
 .github/workflows/               # Scheduled talk-page scraper (not test/build CI)
 docs/                            # Reference docs + design plans (see docs/README.md)
 ```
@@ -138,6 +139,33 @@ WiCE also has no `Source unavailable` rows and ships frozen 2023 Common Crawl ev
 - `OPENAI_API_KEY` - OpenAI
 - `GEMINI_API_KEY` - Gemini
 - `PUBLICAI_API_KEY` - PublicAI models
+
+## Batch Pipeline (`service/`)
+
+A separate, Toolforge-hosted pipeline that runs citation verification unattended, at scale, over a queue of articles rather than one citation a user clicked — the WMF "Edit Suggestions" integration described in `docs/design-plans/2026-08-07-batch-source-checks-for-edit-suggestions.md`. It imports `core/` directly (like `benchmark/` and `cli/` do) and shares no code with `main.js` beyond that.
+
+**Not yet a single runnable pipeline.** Each stage below is built and tested — several against real Wiki Replicas, real Lift Wing calls, and a real ToolsDB write — but stages 1-3 and stages 4-5 have never been run in the same process; see `docs/design-plans/2026-08-24-csv-deliverable-and-component-names.md` for what's missing before `select → fetch → verify → CSV` is one command.
+
+One noun per component; anything runnable is prefixed `run-`:
+
+| Stage | Component | File | Does |
+|---|---|---|---|
+| 1 Select | **Article Picker** | `service/article-picker.js` + `service/replicas.js` | Builds and runs the Wiki Replicas query that picks candidate articles (e.g. pages carrying `{{Failed verification}}`) |
+| 2 Extract | **Claim Extractor** | `service/claim-extractor.js` | Fetches an article's rendered HTML at a pinned revision and pulls out each citation's claim text |
+| 3 Fetch | **Source Fetcher** | `tf-source-fetcher` (separate Toolforge tool), reached via `core/worker.js` | Retrieves the cited source's content. **Stubbed by default** — unattended fetching of third-party URLs from Toolforge isn't yet cleared with WMCS; opt in per-run with `--live-source-fetch` |
+| 4 Verify | **Verifier** | `service/verifier.js` | One model call per claim → verdict, via `core/prompts.js` / `core/providers.js` / `core/parsing.js`. Calls the **LLM Router** (`tf-llm-router`, a separate Toolforge tool) for the model call itself. Halts the whole run (doesn't just log) on a 401/402/403 from the model, so a spent API budget can't silently turn into a queue of wrong "source unavailable" rows |
+| 5 Store | **Finding Builder** + **Findings Store** | `service/finding-builder.js` + `service/findings-store.js` + `service/toolsdb.js` | Turns a verdict into a row (claim hash, TTL, `published:false` pending a precision threshold — see the design doc) and upserts it into ToolsDB's `citation_findings` table |
+| 6 Report | *(not built)* | — | Nothing serves or exports findings yet. The design doc above proposes a CSV writer here as a smaller near-term goal than the read API the parent design (`2026-08-07-...`) originally specified |
+
+Runnable entry points, each a thin wiring layer over the components above:
+
+| Runner | Does | Outbound network |
+|---|---|---|
+| `service/run-pick.js` | Stage 1 only, prints selected articles as JSON | Wiki Replicas |
+| `service/run-extract.js` | Stages 1-3, prints a citations/URLs/fetched/failed funnel | Wiki Replicas, Wikipedia REST, source fetcher (opt-in) |
+| `service/run-replay.js` | Stages 4-5 over `benchmark/dataset.json`'s stored claim/source pairs instead of live fetching — the integration test for verify+store with zero third-party requests | Model API only (+ Wikipedia REST to resolve page IDs) |
+
+No runner yet joins all six stages (`service/run-sweep.js`, proposed, not built).
 
 ## Development Workflow
