+++ service/run-extract.js
@@ -13,15 +13,15 @@
 // after checking WMCS clearance); every other run stays stubbed.
 //
 // Usage (on a Toolforge bastion, inside the tool account):
-//   node service/extract-articles.js
-//   node service/extract-articles.js --criterion citation-needed --max 5
-//   node service/extract-articles.js --live-source-fetch --max 1
+//   node service/run-extract.js
+//   node service/run-extract.js --criterion citation-needed --max 5
+//   node service/run-extract.js --live-source-fetch --max 1
 
 import { JSDOM } from 'jsdom';
 import { parseArgs } from 'node:util';
 import { openReplicaConnection, makeQueryFn } from './replicas.js';
-import { selectCandidates, CRITERIA } from './selection.js';
-import { runBatch, ARTICLE_OUTCOMES } from './pipeline.js';
+import { selectCandidates, CRITERIA } from './article-picker.js';
+import { runBatch, ARTICLE_OUTCOMES } from './claim-extractor.js';
 import { fetchArticleHtml } from '../core/wikipedia.js';
 import { fetchSourceContent } from '../core/worker.js';
 
@@ -52,7 +52,7 @@ function parseCliArgs(argv) {
     };
 }
 
-const HELP_TEXT = `usage: node service/extract-articles.js [options]
+const HELP_TEXT = `usage: node service/run-extract.js [options]
 
 Selects candidate articles from Wiki Replicas, fetches each one's real
 rendered HTML, and extracts its citations and claims. Source fetching (stage
